[X64 08/09 j] Sua sap khi dang nhap lai: sNotify/sSendOp nhan con tro qua int (KMailClient/KAuctionClient/KChienLenhClient, UiAuction/UiChienLenh/UiMail) -> KNPARAM; dump 12:42 KUiMailList::AddHeader dia chi FFFFFFFFFCB9D3F0

+++ Sources/S3Client/Ui/UiCase/UiChatCentre.cpp
@@ -474,7 +474,7 @@ void KUiChatCentre::LoadScheme(class KIniFile* pIni)
 
 	m_GroupBtn[0].GetPosition(&m_nFirstGroupBtnLeft, &m_nFirstGroupBtnTop);
 
-	m_FriendList.SetCustomColorOffset((KNPARAM)&(((KUiPlayerItem*)0)->nData));
+	m_FriendList.SetCustomColorOffset((int)(KNPARAM)&(((KUiPlayerItem*)0)->nData));
 
 	//ºÃÓÑÁÐ±í±³¾°Í¼
 	const char* SectionTitle = "ListBgImg";
@@ -1546,7 +1546,7 @@ void KUiChatCentre::UpdateGroup(int nGroupIndex)
 			}
 
 			m_FriendList.SetContent((unsigned char*)m_pFriendList, m_nNumFriendInList, sizeof(KUiPlayerItem),
-				(KNPARAM)&(((KUiPlayerItem*)0)->Name));
+				(int)(KNPARAM)&(((KUiPlayerItem*)0)->Name));
 		}
 	}
 }
